More work towards a getting a stable conversion
This commit adds handling for compounded entries (Little+Blue+Truck).
It also rewrites these entries so that they conform to the new format
(only first word should be uppercase). The commit also includes work on
handling comments.

The main obstacle, currently, is inserting comments into multi-line tiers.
Right now the script inserts these comments into their correct time
intervals, which might sometimes be inside a multi-line tier. This
causes CHECK to fail, complaining that there should be a delimiter on a
preceding line, since the comment broke the parse-ability of the
multiple lines. By adding a "." after the 0 on the preceding line we
can pass check (while keeping the comment within the multi-line entry).

for example:

*CHF:	0 14233340_14233470
%xcom:	subregion 2 of 5 ends at ............ <-inserted
    &=vfx 14233470_14233840
    0 . 14233840_14233940
%xdb:	average......
*SIL:	0 . 14233940_14300970