Fix interval regex
The interval regex now explicitly looks for the hidden character ("\025") in
order to differentiate it from the adult word count on the line. The
hidden character is then thrown away when the interval value is going
to be used. This is how it should have been done, rather than my stupid
workaround requiring at least 3 digits for the offset.