Found bug in the skeleton/cha merge
This commit fixes some formatting of the comments during merge. I
thought this might solve the merge problem that came up when trying on
new inputs but it didn't. Right now when running on the files for 9_08, a
large proportion of the words that are pulled out of the 9_08.cex
(which I made sure is all of them, implying the original format parsing
is working correctly) are missing in the final 9_08_test_merge.cha.
This is probably due to some word getting skipped (probably due to
consecutive comments or something along those lines) due to the
curr_interval being pushed forward before the appropriate word was
written in, causing the rest of the words in the list to be left out
(since the interval has progressed, but the queue of words are still
waiting with the old interval sitting at the top).