Fix off by one error with intervals and the chun_words function
We can sidestep the intervals being off by +/- 1 millisecond by adding
extra checks for these cases, though this might be a source of
potential bugs in the future. This commit also fixes a bug in the
chunk_words() function which was leaving off the last word from the
input array.

Though the +/- 1 error might be taken care of, after testing on more
files, more discrepancies between the skeleton and .cex file are
showing up. In 12_07.cex, we have interval 10423095_10424260, but in
the .cha skeleton the interval is 10423120_10424260. This doesn't
really look like a rounding error. It's not immediately clear how these
isolated differences are making their way into the .its file to begin
with, while most of the intervals are otherwise in agreement.