Add clan_interval section and fix multi-line parsing bug
The new repository, clan_intervals, checks that the intervals in the
original cex file match those in the annotated version of the same
file. This shows us if there have been any changes to these intervals
in the process of annotating them. The code from the clan_intervals
script is now embeded in the newclan script so that we can make sure
we're dealing with annotated clan files that have intervals identical
to those in the cha skeleton. If we find that these intervals aren't
totally matching, then newclan will stop progressing and tell you
there's a problem (and where in the annotated clan file it was).

This commit changes the parameters that newclan is expecting as input.
The first argument needs to be the path to an unannotated version of
the clan file that we're trying to convert. The rest of the arguments
and their order remains the same.

There was also a bug in the way multi-line entries were being parsed
from the .cex files. This has been fixed.