Ran batch conversion over Subject_Files and found lots of errors
In error_files/ there are reports for all the clan files that had
interval errors in them. This means there are discrepancies
between the intervals in the silences/subregion .cex and the
consensus_final .cex. This has shown that a large proportion of the
files have these discrepancies. For files that were successfully
processed, there is a record of the commands used (which lists what
files were passed into newclan.py) in the batch_processed.txt file
found in this repo.

The last batch process I ran resulted in 110 files being converted. The
files which were not converted either had interval errors, or their
filenames weren't formatted regularly (like not including "final" in
the final .cex filename)